# api/docs/swagger.yaml
basePath: /
definitions:
...
  handler.UserStatAmountRequest:
    properties:
      amount:
        example: 100
        type: integer
    type: object
  handler.UserStatMessageResponse:
    properties:
      message:
        example: Operation completed successfully
        type: string
    type: object
...
  model.UserStat:
    properties:
      experience:
        type: integer
      gold:
        type: integer
      id:
        type: integer
      streak:
        type: integer
      total_plant_harvested:
        type: integer
      total_task_completed:
        type: integer
      updated_at:
        type: string
      user_id:
        type: integer
    type: object
host: 10.155.36.40:8080
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: API для FarmFocus - системы управления задачами и фермой
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: http://swagger.io/terms/
  title: FarmFocus API
  version: "1.0"
paths:
...
  /user-stats:
    get:
      consumes:
      - application/json
      description: Возвращает статистику пользователя. Если статистика не существует
        - создает новую.
      parameters:
      - description: User ID
        in: header
        name: X-User-ID
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.UserStat'
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - ApiKeyAuth: []
      summary: Получить статистику пользователя
      tags:
      - user-stats
  /user-stats/experience:
    post:
      consumes:
      - application/json
      description: Увеличивает количество опыта пользователя на указанное значение
      parameters:
      - description: User ID
        in: header
        name: X-User-ID
        required: true
        type: string
      - description: Количество опыта для добавления
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handler.UserStatAmountRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.UserStatMessageResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - ApiKeyAuth: []
      summary: Добавить опыт пользователю
      tags:
      - user-stats
  /user-stats/gold:
    post:
      consumes:
      - application/json
      description: Увеличивает количество золота пользователя на указанное значение
      parameters:
      - description: User ID
        in: header
        name: X-User-ID
        required: true
        type: string
      - description: Количество золота для добавления
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handler.UserStatAmountRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.UserStatMessageResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - ApiKeyAuth: []
      summary: Добавить золото пользователю
      tags:
      - user-stats
  /user-stats/streak/increment:
    post:
      consumes:
      - application/json
      description: Увеличивает ежедневный стрик пользователя на 1
      parameters:
      - description: User ID
        in: header
        name: X-User-ID
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.UserStatMessageResponse'
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - ApiKeyAuth: []
      summary: Увеличить стрик пользователя
      tags:
      - user-stats
  /user-stats/streak/reset:
    post:
      consumes:
      - application/json
      description: Сбрасывает ежедневный стрик пользователя до 0
      parameters:
      - description: User ID
        in: header
        name: X-User-ID
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.UserStatMessageResponse'
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - ApiKeyAuth: []
      summary: Сбросить стрик пользователя
      tags:
      - user-stats
...
schemes:
- http
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: X-User-ID
    type: apiKey
swagger: "2.0"